Remove circular dependencies with repoData and repoDataSchema. Update storage.h interrace Update storage_ram.c to use the new interface. Update storage_sqlite.c to use the new interface. Remove dependencies. Remove auxiliary functions in repoData.c. Update dat_print_payload_struct to a more memory efficient version.

+++ src/lib/storage/storage_sqlite.c
@@ -19,9 +19,6 @@
  */
 
 #include "storage.h"
-//TODO Remove dependency on repoDataSchema and repoData
-#include "repoDataSchema.h"
-#include "repoData.h"
 #include <sqlite3.h>
 
 ///< Status variables buffer
@@ -31,29 +28,19 @@ static size_t status_entries = 0;
 static char* fp_table = NULL;
 
 ///< Payloads storage buffer
-//static uint8_t *payload_db = NULL;
-//static uint8_t **payloads_sections_addresses = NULL;
 static int payloads_entries = 0;
+static data_map_t *payloads_schema = NULL;
+static char *payloads_table = NULL;
 
 static int storage_is_open = 0;
 
 static sqlite3 *db = NULL;
 
-/**
- * TODO: ADD documentation
- * @param c_type
- * @param buff
- * @param stmt
- * @param j
- */
-void get_sqlite_value(char* c_type, void* buff, sqlite3_stmt* stmt, int j);
-
-/**
- * Translate to sql format a c format type
- * @param c_type
- * @return string in sql syntax
- */
-const char* get_sql_type(char* c_type);
+static void get_sqlite_value(char* c_type, void* buff, sqlite3_stmt* stmt, int j);
+static const char* get_sql_type(char* c_type);
+static int get_payloads_tokens(char** tok_sym, char** tok_var, char* order, char* var_names);
+static void get_value_string(char* ret_string, char* c_type, const char* buff);
+static int get_sizeof_type(char* c_type);
 
 int storage_init(const char *db_name)
 {
@@ -156,14 +143,16 @@ int storage_table_flight_plan_init(char *table, int n_entires, int drop)
     return SCH_ST_OK;
 }
 
-/* TODO: Use the function to allocate each payload table. Data storage have all the info */
-int storage_table_payload_init(char *table, int n_entries, int drop)
+int storage_table_payload_init(char *table, data_map_t *data_map, int n_entries, int drop)
 {
+    if(table == NULL || data_map == NULL)
+        return SCH_ST_ERROR;
+
     int rc = 0;
     if(drop) {
         char *err_msg;
         char *sql;
-        for( int i = 0; i < last_sensor; ++i) {
+        for( int i = 0; i < n_entries; ++i) {
             sql = sqlite3_mprintf("DROP TABLE IF EXISTS %s",  data_map[i].table);
             rc = sqlite3_exec(db, sql, 0, 0, &err_msg);
             if (rc != SQLITE_OK ) {
@@ -174,7 +163,7 @@ int storage_table_payload_init(char *table, int n_entries, int drop)
         }
     }
 
-    for(int i=0; i < last_sensor; ++i) {
+    for(int i=0; i < n_entries; ++i) {
         char create_table[SCH_BUFF_MAX_LEN * 4];
         memset(&create_table, 0, SCH_BUFF_MAX_LEN * 4);
         snprintf(create_table, SCH_BUFF_MAX_LEN * 4, "CREATE TABLE IF NOT EXISTS %s(id INTEGER, tstz TIMESTAMPTZ,",
@@ -186,7 +175,7 @@ int storage_table_payload_init(char *table, int n_entries, int drop)
         char var_names[SCH_BUFF_MAX_LEN * 4];
         memset(&var_names, 0, SCH_BUFF_MAX_LEN * 4);
         strcpy(var_names, data_map[i].var_names);
-        int nparams = get_payloads_tokens(tok_sym, tok_var, order, var_names, i);
+        int nparams = get_payloads_tokens(tok_sym, tok_var, order, var_names);
 
         for (int j = 0; j < nparams; ++j) {
             char line[100];
@@ -206,6 +195,11 @@ int storage_table_payload_init(char *table, int n_entries, int drop)
             return SCH_ST_ERROR;
         }
     }
+
+    payloads_entries = n_entries;
+    payloads_table = strdup(table);
+    payloads_schema = data_map;
+
     return SCH_ST_OK;
 }
 
@@ -497,18 +491,76 @@ void get_sqlite_value(char* c_type, void* buff, sqlite3_stmt* stmt, int j)
     }
 }
 
-
-int storage_payload_set_data(int payload, int index, void *data, size_t size)
+int get_payloads_tokens(char** tok_sym, char** tok_var, char* order, char* var_names)
 {
-    if(payload >= last_sensor) return SCH_ST_ERROR;
+    const char s[2] = " ";
+    tok_sym[0] = strtok(order, s);
+
+    int j=0;
+    while(tok_sym[j] != NULL) {
+        j++;
+        tok_sym[j] = strtok(NULL, s);
+    }
+
+    tok_var[0] = strtok(var_names, s);
+
+    j=0;
+    while(tok_var[j] != NULL) {
+        j++;
+        tok_var[j] = strtok(NULL, s);
+    }
+    return j;
+}
+
+void get_value_string(char* ret_string, char* c_type, const char* buff)
+{
+    if(strcmp(c_type, "%f") == 0) {
+        if (*((int *)buff) == -1) {
+            sprintf(ret_string, " 'nan'");
+        } else {
+            sprintf(ret_string, " %f", *((float*)buff));
+        }
+    }
+    else if(strcmp(c_type, "%d") == 0) {
+        sprintf(ret_string, " %d", *((int*)buff));
+    }
+    else if(strcmp(c_type, "%u") == 0) {
+        sprintf(ret_string, " %u", *((unsigned int*)buff));
+    } else if(strcmp(c_type, "%i") == 0) {
+        sprintf(ret_string," %i", *((int*)buff));
+    }
+}
+
+int get_sizeof_type(char* c_type)
+{
+    if(strcmp(c_type, "%f") == 0) {
+        return sizeof(float);
+    }
+    else if(strcmp(c_type, "%d") == 0) {
+        return sizeof(int);
+    } else if(strcmp(c_type, "%u") == 0) {
+        return sizeof(int);
+    } else if(strcmp(c_type, "%i") == 0) {
+        return sizeof(int);
+    }
+    else {
+        return -1;
+    }
+}
+
+
+int storage_payload_set_data(int payload, int index, void *data, data_map_t *schema)
+{
+    if(data == NULL || schema == NULL)
+        return SCH_ST_ERROR;
 
     char* tok_sym[300];
     char* tok_var[300];
     char *order = (char *)malloc(300);
-    strcpy(order, data_map[payload].data_order);
+    strcpy(order, schema->data_order);
     char *var_names = (char *)malloc(1000);
-    strcpy(var_names, data_map[payload].var_names);
-    int nparams = get_payloads_tokens(tok_sym, tok_var, order, var_names, payload);
+    strcpy(var_names, schema->var_names);
+    int nparams = get_payloads_tokens(tok_sym, tok_var, order, var_names);
 
     char *values = (char *)malloc(1000);
     char *names = (char *)malloc(1000);
@@ -537,7 +589,7 @@ int storage_payload_set_data(int payload, int index, void *data, size_t size)
     strcat(names, ")");
     strcat(values, ")");
     char*  insert_row = (char *)malloc(2000);
-    sprintf(insert_row, "INSERT INTO %s %s VALUES %s",data_map[payload].table, names, values);
+    sprintf(insert_row, "INSERT INTO %s %s VALUES %s", schema->table , names, values);
     free(order);
     free(var_names);
     free(values);
@@ -554,17 +606,18 @@ int storage_payload_set_data(int payload, int index, void *data, size_t size)
     return SCH_ST_OK;
 }
 
-int storage_payload_get_data(int payload, int index, void *data, size_t size)
+int storage_payload_get_data(int index, void *data, data_map_t *schema)
 {
-    if(payload >= last_sensor) return SCH_ST_ERROR;
+    if(data == NULL || schema == NULL)
+        return SCH_ST_ERROR;
 
     char* tok_sym[300];
     char* tok_var[300];
     char order[300];
-    strcpy(order, data_map[payload].data_order);
+    strcpy(order, schema->data_order);
     char var_names[1000];
-    strcpy(var_names, data_map[payload].var_names);
-    int nparams = get_payloads_tokens(tok_sym, tok_var, order, var_names, payload);
+    strcpy(var_names, schema->var_names);
+    int nparams = get_payloads_tokens(tok_sym, tok_var, order, var_names);
 
     char values[1000];
     char names[1000];
@@ -583,8 +636,7 @@ int storage_payload_get_data(int payload, int index, void *data, size_t size)
     }
 
     char get_value[2000];
-    sprintf(get_value,"SELECT %s FROM %s WHERE id=%d LIMIT 1"
-            ,names, data_map[payload].table, index);
+    sprintf(get_value,"SELECT %s FROM %s WHERE id=%d LIMIT 1", names, schema->table, index);
 
     char* err_msg;
     int rc;
@@ -614,7 +666,7 @@ int storage_payload_get_data(int payload, int index, void *data, size_t size)
 
 int storage_payload_reset(void)
 {
-    return storage_table_payload_init(NULL, payloads_entries, 1);
+    return storage_table_payload_init(payloads_table, payloads_schema, payloads_entries, 1);
 }
 
 int storage_payload_reset_table(int payload)
